test: nftTransfer and add sample response
